Hide strict map modules with containers 0.5.
This is a bit of a hack, but it makes it relatively easy to support
both containers < 0.5 and containers 0.5, and depend on
strict-containers. Previously, the module names clashed with
containers 0.5. By hiding them here, you import either the one from
containers or the one from strict-containers, depending on what's
available.